[api] Widen snapshots pit of success (#4392)
This PR makes the snapshot loader preserve page states that are not
explicitly overridden so that, e.g. you can load a previous snapshot
version of a document without switching page or resetting camera simply
by omitting the session state option.

The goal is to provide a nice and robust version of the solution
mentioned in #4381

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [x] `api`
- [ ] `other`

### Test plan

- [x] Unit tests


### Release notes

- Improved loadSnapshot to preserve page state like camera position and
current page if no session snapshot is provided.