inline nanoid (#4410)
We have a bunch of code working around the fact that nanoId is only
distributed as an ES module, but we run both as es and commonjs modules.
luckily, nanoid is nano! and even more so if you strip out the bits we
don't use. This replaces the nanoid library with a vendored version of
just the part we use.

### Change type

- [x] `improvement`