faster (& more!) export snapshot tests (#4411)
Our playwright snapshot tests take quite a long time because there's so
many of them. For each one we have to open a page, load some content,
generate the snapshot, and compare it to what we have stored.

This is pretty time consuming, but a lot of the overhead in these steps
is more or less constant, regardless of the size of the snapshot. This
diff merges our individual snapshot tests into a smaller number of giant
snapshots containing many test cases. It also adds some new tests that
we didn't have before (images, geo shapes, frames).

This revealed that right now, image snapshots are broken - they don't
respect flip X/Y. I'll fix that in a follow-up.

### Change type

- [x] `other`

---------

Co-authored-by: huppy-bot[bot] <128400622+huppy-bot[bot]@users.noreply.github.com>