Fix an issue with canceling the page name and presence name changes. (#4408)
You can now use escape to cancel the changes to page name and presence
name changes.

There was an issue with using Escape to cancel the page name or presence
name changes. The problem was that Escape key event (on key down) was
captured by the popover component (radix) which closed the popover. This
meant that the `onKeyUp` in `TldrawUiInput` never triggered when Escape
was pressed so the `onCancel` callback was never called.
We now prevent the direct closing of the popover on Escape which means
that we do have to close it manually.

Could maybe also cause the `onCancel` to be called on key down, might be
better?

### Before


https://github.com/user-attachments/assets/151a980e-0a24-44cb-a1eb-89e384883ef9

### After

https://github.com/user-attachments/assets/601a5830-1ea8-4849-b489-e2077e746e97



### Change type

- [x] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Create a shape...
2.

- [ ] Unit tests
- [ ] End to end tests

### Release notes

- Fix an issue with not being able to cancel out changing of page names
and user presence names.