Add sleep fn (#4454)
There are a bunch of places we do something like `await new
Promise(resolve => setTimeout(resolve, 1000))`. Most of the time, these
places have to come with an eslint disable. We have a rule enforcing
that timers come from `editor.timers`, but in those cases this won't do
what you would expect: when the editor is disposed, those timers will
never resolve, which isn't really what you want if you're using the
timers for control-flow around more expensive resources that need to be
disposed of.

For example:
```tsx
const thing = createAnExpensiveResourceThatNeedsToBeExplicitlyDestroyed()
try {
    await doSomethingWithIt(thing)
    await new Promise(r => editor.timers.setTimeout(r, 100))
} finally {
    thing.dispose()
}
```

In this example, if the editor is disposed, then the `setTimeout`
callback is never called, and the promise never resolves. The code in
the finally block never runs, and the memory for `thing` leaks.

### Change type

- [x] `other`

### Release notes

(internal-only change)