File save warning (#4469)
This PR adds a one-time warning about files in tldraw. It's only shown
when the user loads a project with >500 records in the store. Hopefully
this cuts down on the number of people who accidentally delete their
local project. It is only shown once and will be dismissed once closed.

<img width="1267" alt="image"
src="https://github.com/user-attachments/assets/c6cac01d-34e3-4b70-a0d6-a4dfcbbdc56d">

We could also pair this with a "local" chip in the top zone that pops up
a screen like this.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- Adds save warning.