Better docs search (#4485)
Reworks our docs search.
* Long pages are now split up into many search entries according to
their headings
* The full content of pages is included in the search index
* Fine tune the ranking til it feels good

I also removed some old search stuff, and ended up doing a bit of a
drive-by refactor of `ContentDatabase`.

The bulk of the work here was actually extracting the heading and the
plain-text content within them to populate the search index. Check
`parse-markdown.tsx` for that.

### Change type

- [x] `other`

---------

Co-authored-by: Mime Čuvalo <mimecuvalo@gmail.com>