Fix an issue with the dirty state of flags not working after the first change to a file (#4564)
We had this logic to prevent marking a file dirty as soon as you opened
it. Looks like this is not longer needed and it actually prevented the
file to get correctly flagged as dirty when you made the first change to
it.

After speaking to @ds300 it seems likely that this is due to a change to
how loading of snapshots works - it no longer triggers the store.listen
callbacks.

### Before:

1. Open new tldraw file.
2. Create a shape.
3. File would not be marked as dirty.


https://github.com/user-attachments/assets/20eb26b2-1542-4d47-8583-ff5be72df524

### Now:

1. Open new tldraw file.
2. Create a shape.
3. File is now correctly marked as dirty.


https://github.com/user-attachments/assets/fba20ed0-3011-4981-b63e-e265bc23a478

Fixes https://github.com/tldraw/tldraw/issues/3940

### Change type

- [x] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- Fixed a bug with the first change after opening a file not marking the
file as dirty.