[sync] Allow doing CRUD directly on the server (#4559)
We've had a couple of requests for the ability to update the room data
directly on the server, which totally make sense when implementing any
kind of workflow that involves talking to an external service.

This PR adds a simple helper that gives you an async transactional
context in which to perform updates that will be committed at the end,
and which are isolated from any other concurrent updates happening in
the server due to client activity or even other transactions. At the end
of a transaction, the generated diff is calculated as if it were on a
client and 'sent' to the room via a push request.

### Change type

- [x] `feature`
- [x] `api`

### Test plan

- [x] Unit tests
- [ ] End to end tests

### Release notes

- Adds the `updateStore` method to the `TLSocketRoom` class, to allow
updating room data directly on the server.