Don't upload a snapshot from the snapshots page (#4577)
This prevents uploading a snapshot when we are on a snapshot page. We
can just use the current link instead of generating a new one.

Pulled out a component for showing the Copy snapshot link share menu
group. Also made the UX consistent (for example, on the share menu we
didn't show the checkmark when copying the link).

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Share a snapshot.
2. Open the snapshot.
3. Copy the share link. This should not upload the current snapshot to
the server and should result copy the current url to the clipboard.

### Release notes

- Copying a snapshot link from a snapshot page now just returns the
current url.