[feature] isShapeHidden option (#4446)
This PR adds an option to the Editor that allows people to control the
visibility of shapes. This has been requested a couple of times for
different use-cases:

- A layer panel with a visibility toggle per shape
- A kind-of 'private' drawing mode in a multiplayer app.

So to test this feature out I've implemented both of those in minimal
ways as examples.

### Change type

- [x] `feature`


### Test plan

- [x] Unit tests


### Release notes

- Adds an `isShapeHidden` option, which allows you to provide custom
logic to decide whether or not a shape should be shown on the canvas.