Support auto switching wheel behavior
This implements support for setting wheelBehavior to auto. What this
means is that it will try to detect if you're using a mouse or a
touchpad, and use wheelBehavior zoom when a mouse is used and pan when a
touchpad is used. This is more intuitive behavior than always having it
set to wheel or pan in my opinion.

As far as I know, it's unfortunately not possible to know if wheel
events are generated from a mouse or a touchpad. The heuristic
implemented is that if you scroll horizontally it assumes you use a
touchpad and switch to pan, and if you pan with the middle mouse button
it assumes you're using a mouse and switches to zoom. This seems to be
similar to what other similar software does.

Fixes #4579