Fix VS Code `editorPath`
I'm not sure why the previous code was written the way it was, but it
causes an error on my system that prevents the extension from running in
development mode:

```
/home/vladh/sources/tldraw/apps/vscode/extension/dist
(9) ['', 'home', 'vladh', 'sources', 'tldraw', 'apps', 'vscode', 'extension', 'dist']
(7) ['home', 'vladh', 'sources', 'tldraw', 'apps', 'vscode', 'extension']
Error: ENOENT: no such file or directory, watch 'home/vladh/sources/tldraw/apps/vscode/extension/editor/index.js'
```

This is because the previous code was removing the leading slash from
paths, which led the watcher to always fail. My solution fixes this, and
does in a much simpler way what I believe the original code was trying
to do, which is to remove the last path component.