text: followup fix to locking text shapes (#4632)
followup to https://github.com/tldraw/tldraw/pull/4569
Once the text shape is locked you can't click the Select tool to switch
out. This addresses that (because the tool is masked at the time). If
there's a better fix for this, I'm all ears, I wasn't sure how else to
resolve this (I see you pulled it out in the other PR @steveruizok )

### Change type

- [x] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- Fix bug with text shape locking.

---------

Co-authored-by: Steve Ruiz <steveruizok@gmail.com>