selection: allow cmd/ctrl to add to selection (#4570)
In the How-To sesh, I noticed that using Shift of course lets you add to
a selection of shapes, but Cmd/Ctrl does not.
Typically, cmd/ctrl lets you do this in other contexts so some of that
muscle memory doesn't get allowed in tldraw currently.
This enables cmd/ctrl to have the same behavior as shift.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- Selection: allow cmd/ctrl to add multiple shapes to the selection.

---------

Co-authored-by: Steve Ruiz <steveruizok@gmail.com>