Fix an issue with nearest point and lines that start and end at the same point (#4650)
This fixes an issue with getting the nearest point to a line for lines
that start and end at the same point.

Not sure how those lines got created though. Programatically we seem to
allow that, so I also added a `onBeforeHandler` to line shape util that
will nudge the end point just slightly if we try to create a line like
that. We could also avoid creating it completely?

[Example sentry
report](https://tldraw.sentry.io/issues/5936469482/?alert_rule_id=12855294&alert_timestamp=1727804797951&alert_type=email&environment=production&notification_uuid=3482b5d8-ad95-48ca-be09-40c77af5fcf2&project=4504203639193600&referrer=alert_email)
for this issue.

### Change type

- [x] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Create a line shape programatically and make the start and end point
be the same. Think it has to start and end at (0,0) but not sure.
2. Brush select.
3. This used to throw an error since we could not get the distance to
line, but that should no longer be happening.

### Release notes

- Fix a bug with nearest points for lines that start and end at the same
point.