Arrowhead clipping fix (#4646)
Fixes issues with arrowheads not correctly clipping the arrow's body.
The main issue was when we used the fill none option. There was also a
difference in how it worked if arrows had labels vs if they didn't.

Also transitions the arrow indicator form using masks to using clip
paths so that we also improve the performance of panning when we have
arrows selected.

[Related
discussion](https://discord.com/channels/859816885297741824/1290288009492824065/1290906937797640273).

### Before


![image](https://github.com/user-attachments/assets/1669fc2a-e3a3-4496-ab19-2a48bd6c8bee)

### After


![image](https://github.com/user-attachments/assets/c191e83c-b8c9-4974-a599-71bc49a2490b)

### Change type

- [x] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- Fix an issue introduced with #4636. The arrowheads did not correctly
clip the arrow body.

---------

Co-authored-by: huppy-bot[bot] <128400622+huppy-bot[bot]@users.noreply.github.com>
Co-authored-by: Steve Ruiz <steveruizok@gmail.com>