dotcom top bar / .tldr file drops (#4661)
This PR adds a few somewhat random features to dotcom and tldraw:

### New botcom header

This PR removes the header from botcom and places the title and page
menu on the canvas. It adds a temporary menu that we can modify later to
include other items from the menu. I'm actually tempted to bring back
the help menu for some of these.

### .tldr frile dropping

This PR adds support for dropping tldraw files onto the side menu,
though it does not implement actual file uploading.

### Change type

- [x] `internal`