[sync] Set instance.isReadonly automatically  (#4673)
Follow up to #4648 , extracted from #4660 

This PR adds a TLStore prop that contains a signal for setting the
readonly mode. This allows the readonlyness to change on the fly, which
is necessary for botcom. it's also just nice for tlsync users to be able
to decide on the server whether something is readonly.

### Change type

- [x] `improvement`

### Release notes

- Puts the editor into readonly mode automatically when the tlsync
server responds in readonly mode.
- Adds the `editor.getIsReadonly()` method.
- Fixes a bug where arrow labels could be edited in readonly mode.