Add missing CSS imports to examples (#4689)
We include the `import 'tldraw/tldraw.css'` import to make our examples
more copy and pasteable, but we don't test for this so sometimes it gets
missed. Casual update of a few examples to include the CSS import (and
to add Example to the filename to make them easier to grep).

### Change type

- [x] `other`