Lod Perf (effect scheduler) (#4682)
This PR aims to improve on the LOD performance of image and video
elements. Its main goal is to reduce the number of renders caused by the
hooks in `useAsset`. It does this by moving the majority of the logic
into an effect scheduler (via our `react`), keeping the result (url and
asset) in React state, and only updating the result it has actually
changed. We also debounce subsequent renders

### Change type

- [x] `improvement`

### Release notes

- The `useAssets` hook now is now specifically for `TLImageAsset`- or
`TLVideoAsset`-type assets.