[botcom] local session state, logged out view of files (#4711)
This PR implements a few changes inspired by the instantdb branch.

## Providers

I've flattened the component tree for react context providers in
`TlaProviders`, which were getting complex. This should make it easier
in the future to do clean diffs as we adjust providers. I started it,
but we should avoid having too many components that return other
wrappers and providers in this file.

## Local session state

The `TldrawAppSessionsState` is ignored. I have not removed it from the
schema, as I don't want to handle that right now, but we no longer read
it or update it. Instead, we keep a local atom that tracks data that was
previously in the session record: ie sidebars, theme, auth user id, etc.

## Removing dependencies on `app`

The change to `localSessionStorage` removes several places where we
depended on the `app`, as well as all of the places where we updated it.

## Offline preferences for export settings

Export settings are currently sitting on the `TldrawAppUser` record. In
this PR, they also are at `localSessionStorage.exportSetttings` as a
backup for signed-out users.

## Offline editor changes

When visiting a file while logged out, the editor now displays the
`AnonLayout`, i.e. a framing around the editor.


![image](https://github.com/user-attachments/assets/9c607fac-a376-412a-afe2-c93c971d58ef)

### People menu

When visiting another user's file while logged out, I've included the
`PeopleMenu` in a location next to the Style Panel.


![image](https://github.com/user-attachments/assets/d173d4d3-1ce6-4144-beb0-b47ec922eae8)

### Export menu

For logged out users, I've (perhaps temporarily) replaced the toggle
sidebar button with a button that opens an export menu.


![image](https://github.com/user-attachments/assets/f6ae4856-9377-4acb-928c-ed8558b155b8)


### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [x] `other`