menus: rework the open menu logic to be in one consistent place (#4642)
We have a lot of logic scattered everywhere to prevent certain logic
when menus are open. It's a very manual process, easy to forget about
when adding new shapes/tools/logic. This flips the logic a bit to be
handled in one place vs. various places trying to account for this.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- Rework open menu logic to be centralized.

---------

Co-authored-by: Steve Ruiz <steveruizok@gmail.com>