Fix bemo deploy on publish-new (#4721)
Our publish-new script was not allowing the bemo deploys to go out
because it was pushing the changelog commits to the `bemo-production`
branch, and the changelog commits have `[skip ci]` in them which causes
the bemo deploy script to not run.

It was working fine for patch releases, and I guess the reason we only
noticed this recently is because we shipped our first
forwards-incompatible backend change in aaages last week.

### Change type

- [x] `other`