Add a camera config option `autoWheelBehavior`
When `wheelBehavior` is set to `'auto'`, this controls whether the
behavior used is `'pan'` or `'zoom'`. The value will automatically be
changed on certain user interactions. Specifically it will change to
`'pan'` when scrolling horizontally and change to `'zoom'` when the
canvas is panned by either spacebar dragging or middle mouse dragging.