[botcom] Publishing (#4688)
Adds publishing to botcom. 

This allows the users to publish the existing document. This is a point
in time snapshot, which they can then update or delete at a later time.
Only the owner of the file has the permission to do that, while everyone
with a link can view the published document.

### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [x] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- Add publishing to botcom.

---------

Co-authored-by: David Sheldrick <d.j.sheldrick@gmail.com>
Co-authored-by: Steve Ruiz <steveruizok@gmail.com>