[botcom] file state (#4766)
This PR

- Replaces the file-edit and file-view records with a single record
called `file-state` that has timestamps for first view and most recent
edit.
- Uses the above to simplify the recent files sorting (maintains the
same behaviour, just achieves it with simpler code)
- removes unused presence stuff (should have been removed back in that
burn-it-all-down pr after we merged the prototype), we can add the UI
bits back when we work on file-level presence.
- Stores the UI state in the db so you go back to where you were before
in the file next time you open the file

implements INT-342

### Change type

- [x] `other`