[botcom] Duplicate / Publish / Create / Delete files on the server  (#4798)
This PR improves and creates functionality on the backend.

For the most part, simple changes such as creating a file or changing
its name rely on optimistic updates. However, some actions rely on
either sending information to the server directly. While we can make
optimistic updates in these situations, we also need to revert those
changes if the request fails.

## Duplicating files

This PR adds duplication on the server. Previously, the duplicate
feature required that the duplicating file be open, so duplicating other
files (from the sidebar) would duplicate the current open file instead.
Now the server will pull the latest file from the tldraw worker and
duplicate that instead.

If the user is viewing the duplicated page, we redirect them to the new
page.

## Deleting / forgetting files

This PR adds deletion on the server. If the user owns a file, they will
optimistically delete the file and all states associated with the file;
if the user does not, then they'll only "forget" the file by
optimistically deleting all of their own states associated with the
file. Meanwhile, the server will actually delete the file and the file's
published snapshot. If the delete fails, the user's optimistic updates
will be reversed.

If the user is viewing the deleted / forgotten file, we redirect them to
the root.

## Publishing / un-publishing files

This PR adds publishing on the server. Previously, the user would upload
the current editor state to the server. Now, the server pulls the latest
file from the tldraw worker and creates a published snapshot based on
that instead.

I'd still like to build the "publish history" of versions that a user
has published for a given file, rather than overwriting the single
published snapshot file.

## Creating files

This PR adds support for .tldr file drops. Files are created on the
server and then added to the user's own files.

## Cleanup

The sync-worker was getting pretty messy so I did some cleanup. Mostly,
I've moved the endpoints that are related to the app to a `tla` folder,
as I've done elsewhere. It wasn't clear to me which endpoints were
"purely" working for the app layer and which weren't.

### Change type

- [x] `other`

---------

Co-authored-by: David Sheldrick <d.j.sheldrick@gmail.com>