[botcom] improve error UX (#4790)
This PR makes it so that

- if you visit a URL for a private file and you are signed out, you see
this
<img width="546" alt="image"
src="https://github.com/user-attachments/assets/1cd546e2-6b95-478f-be8e-51166635de1c">
Before this PR you'd be redirected immediately to a sign in page, which
is disorienting and not necessarily what the user should have to do.
- If you visit a URL for a private file and you are signed in, you see
this.
<img width="1137" alt="image"
src="https://github.com/user-attachments/assets/39bd8f5c-ad31-4571-b923-be4a766d3e20">
Before this PR you would see a flash of the sidebar and then it'd
immediately disappear, after which you'd only see the error message with
a 'return to tldraw' link that took you back to `/`.

In order to achieve all this I had to do some minor refactorings which
I'll write up on monday.



### Change type


- [x] `other`