Better support scale / quality in export utilities (#4795)
This PR allows for better control over scale when using the image export
utilities in tldraw.

- `scale` is used for just the output scale of an image, relative to the
size in tldraw
- `pixelRatio` is used for the pixelRatio scale, default is 2

A shape that is 100x100 in tldraw:
- with { scale: 1 } should export as a 100x100 svg
- with { scale: 2 } should export as a 200x200 svg
- with { scale: 1, pixelRatio: 1 } should export as a 100x100 png at
72dpi
- with { scale: 1, pixelRatio: 2 } should export as a 200x200 png at
144dpi
- with { scale: 2, pixelRatio: 2 } should export as a 400x400 png at
144dpi

Note, I tried deprecating scale instead of renaming to avoid the
breaking change, but I don't think we can get away with that since it
gets spread in from other places (where scale is used for actual 1x 2x
scale).

### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [x] `api`
- [ ] `other`

### Test plan

1. Export an image at scale 1.
2. Export an image at scale 2.
2. Export an image at scale 3.

### Release notes

- Improved treatment of `scale` in image copy / export utilities.