Only send updates to active users (#4894)
This adds user durable object registration so that we can track which
users are active and only send updates to them.

The way it works:
1. When the user first connects we register register the user with the
replicator.
2. The replicator stores the data about active users in a new
`active_user` table.
3. When replicator receives postgres changes it now also looks into the
`active_user` table to get active impacted user ids.
4. It then forwards these updates to user DO. These forwards succeed
even if the user's DO was not active before. We can use the user DO's
state to check for whether it has active users connected. If it doesn't
then we unsubscribe it. This means that any further updates won't be
sent to it.

Resolves INT-459

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- Only send updates to active users.

---------

Co-authored-by: David Sheldrick <d.j.sheldrick@gmail.com>