make sure copy-as-png comes in at natural size (#4771)
Browsers sanitize image formats to prevent security issues when pasting
between applications. For
paste within an application though, some browsers (only chromium-based
browsers as of Nov 2024)
support custom clipboard formats starting with "web " which are
unsanitized. Our PNGs include a
special chunk which indicates they're at 2x resolution, but that
normally gets stripped - so if
you copy as png from tldraw, then paste back in, the resulting image
will be 2x the expected
size. 

To work around this, this diff writes 2 version of the image to the
clipboard - the normal png, and
the same blob with a custom mime type. When pasting, we check first for
the custom mime type, and
if it's there, use that instead of the normal png.

### Change type

- [x] `bugfix`

### Release notes

- Shapes copied as PNG will have the same size when pasted back into
tldraw.