handle postgres failures in replicator (#4912)
- each time the replicator reboots it gets a new random `bootId`
- each time the replicator sends a row update message to a user DO it
includes the boot id.
- if the boot id doesn't match the previous one, the user DO resets its
state and makes the sockets reconnect to ensure consistent data.
- each time the replicator's postgres subscription fails it tries to
reboot iself
- active user DOs ping the replicator every 10 seconds to make sure it
wakes up if it was put to sleep for whatever reason (I think this is
unlikely to happen, but cloudflare is technically allowed to stop DOs
for whatever reason and they won't start up again unless we perform a
fetch or RPC on them. Better safe than sorry)
- no longer using blockConcurrencyWhile in replicator init. Instead we
have a mechanism for the replicator RPCs (fetchDataForUser et al) to
wait for a 'connected' state to be achieved.
- gonna add a ticket to add timeouts for these so that we can handle
situations of low/zero availability more gracefully for end users.

also

- don't throw away the active_user table on sqlite init. This could lead
to zombie user DOs that are active but don't receive updates.


### Change type

- [x] `other`