Fix file name when exporting a single unnamed frame
If you have a frame without a name, select it and export it, the name
before the extension would end up as an empty string. So the final name
would just be a period and the extension (e.g. `.svg`). This would cause
the SVG file to be downloaded as `svg.txt` with Chrome and as just `svg`
with Firefox (and equivalent for JSON files, but PNG files would be
named `png.png`).

With this change they will be named `frame.<extension>` instead.