Error link blue (#4932)
Describe what your pull request does. If you can, add GIFs or images
showing the before and after of your change.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [x] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. `yarn run dev`
2. Develop...
3. Check by importing the component somewhere in the example.(dev page.)

- [ ] Unit tests
- [ ] End to end tests

### Release notes
close #4927 


![image](https://github.com/user-attachments/assets/01b1a73f-d26b-456f-8617-e7236f028fe4)

```
<a href={url.toString()} style={{ color: 'blue', textDecoration: 'underline' }}>
	create a GitHub issue
</a>
```
Added blue underline style to links in error messages.
Currently, tldraw seems to be styling with CSS. If it's good, I'll add a
style for that function in css.
If so, more styles can be applied.
For ex...(There are concerns that writing this as style props may reduce
readability.)
```
a {
  color: #0000ee; 
  text-decoration: underline; 
  cursor: pointer; 
}

a:visited {
  color: #551a8b; 
}

a:hover {
  color: #0000cc; 
  text-decoration: underline; 
}

a:active {
  color: #ff0000; 
}
```

---------

Co-authored-by: Steve Ruiz <steveruizok@gmail.com>